Cleanup in scripts, add add logo for gh, rename main cmd package
