--- internal/gcpterraform/README.md
@@ -1,6 +1,0 @@
-# OUTDATED INFO
-How to build project: 
-
-go get -d
-go build
-./gcp-terraform
