--- internal/gcpterraform/client_private.key
@@ -1,1 +1,0 @@
-aGsGuo9ODki0ZpS1U3c28tsI6UWjCW1Gbn8lIYRamXA=
