--- internal/gcpterraform/client_public.key
@@ -1,1 +1,0 @@
-KQwNg8z7nSgD23nHga8PKeSrh2GupEstDkQ3Jww5eg4=
