--- internal/gcpterraform/infrastructure/sysctl.conf
@@ -1,7 +1,0 @@
-net.ipv4.ip_forward = 1
-net.ipv6.conf.default.forwarding = 1
-net.ipv6.conf.all.forwarding = 1
-net.ipv4.conf.all.rp_filter = 1
-net.ipv4.conf.default.proxy_arp = 0
-net.ipv4.conf.default.send_redirects = 1
-net.ipv4.conf.all.send_redirects = 0
