--- internal/gcpterraform/server_private.key
@@ -1,1 +1,0 @@
-SDZnNuMWQz+cKlr6f7Vu+Q98R+sl1D9EJPmDWWJZaUM=
