--- internal/gcpterraform/server_public.key
@@ -1,1 +1,0 @@
-G8naBU85RGmh2iZBi2KL3qomJOGKy5jvU97bO2I5tQ4=
