--- internal/gcpterraform/vm-packages
@@ -1,20 +1,0 @@
-sudo yum install git wget
-
-
-sudo yum-config-manager --setopt=centosplus.includepkgs=kernel-plus --enablerepo=centosplus --save
-sudo sed -e 's/^DEFAULTKERNEL=kernel$/DEFAULTKERNEL=kernel-plus/' -i /etc/sysconfig/kernel
-sudo yum install kernel-plus wireguard-tools
-sudo reboot
-
-
-sudo yum install epel-release elrepo-release
-sudo yum install yum-plugin-elrepo
-sudo yum install kmod-wireguard wireguard-tools
-
-
-
- sudo yum install -y yum-utils
-
-  sudo yum-config-manager \
-    --add-repo \
-    https://download.docker.com/linux/centos/docker-ce.repo
