+++ scripts/README.md
@@ -1,0 +1,6 @@
+There are some notes and scripts _for development purpose only_
+
+<figure>
+    <img alt="There must be a funny running jumping gopher" src="https://github.com/egonelbre/gophers/raw/10cc13c5e29555ec23f689dc985c157a8d4692ab/.thumb/animation/2bit-sprite/demo.gif ">
+    <figcaption>Kudos to <a href="https://github.com/egonelbre/gophers">egonelbre</a></figcaption>
+</figure>
