+++ scripts/firewall
@@ -1,0 +1,3 @@
+ firewall-cmd --permanent --zone=public --add-port=35053/udp
+ firewall-cmd --permanent --zone=public --add-masquerade
+ firewall-cmd --reload
