+++ scripts/wireguard-installation.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+sudo yum install epel-release elrepo-release -y
+sudo yum install yum-plugin-elrepo -y
+sudo yum install kmod-wireguard wireguard-tools -y
