--- testdata/client.conf
@@ -1,9 +1,0 @@
-[Interface]
-Address = 10.13.13.6/32
-PrivateKey = eEvSjfU/nyt1U1tC0Sx28eWtRIFbaxyICOl9Gvr3VH0=
-DNS = 8.8.8.8, 8.8.4.4
-
-[Peer]
-PublicKey = 83Qm11F86icBjd7eHukUQyBN2JqzKvJe2+AShF4/gkY=
-AllowedIPs = 0.0.0.0/0
-Endpoint = 34.91.35.38:51820
